chore(api): move stage acquisition models from shared to prescription context
