+++ api/src/prescription/campaign-participation/domain/models/StageAcquisitionCollection.js
