+++ api/src/prescription/campaign-participation/domain/models/StageCollection.js
