+++ api/tests/tooling/domain-builder/factory/user-campaign-results/build-stage-collection.js
@@ -1,4 +1,4 @@
-import { StageCollection } from '../../../../../src/shared/domain/models/user-campaign-results/StageCollection.js';
+import { StageCollection } from '../../../../../src/prescription/campaign-participation/domain/models/StageCollection.js';
 
 const buildStageCollection = function ({ campaignId, stages } = {}) {
   return new StageCollection({ campaignId, stages });
